requirements.txt
aiohttp
aiosignal
aiosqlite
async-timeout
click
colorlog
colorama
covid
dnspython==2.3.0
environs
ffmpeg-python
frozenlist
GitPython
googletrans
gTTS
h11
h2
hpack
httpcore==1.0.2
httpx
Pillow>=10
pydantic>=1.10, <2
pymongo
Pyrogram
python-dotenv
qrcode
requests
TgCrypto
yt-dlp
google-generativeai